Use SafeLoader, since Loader is a alias for UnsafeLoader
See https://msg.pyyaml.org/load

As this is used to load file that are stored along the git checkout,
a attacker could inject a _clc.yaml or _clc_history.yaml that would
erase the one stored in the directory, and then be able to execute
code since the default loader is not the safe one.

The CLoader shouldn't be affected by the problem.